Add files via upload
PACE Inc., a diversified corporation based in Canada, owns a subsidiary courier company situated in Brazil. The organization has provided the dataset on absenteeism for analysis, with the overarching objective of comprehending absenteeism and its contributing factors within their operational context.

This report analyzes absenteeism at the courier company from July 2007 to July 2010. It provides insights into various factors contributing to absenteeism, such as reasons for absence, and health-related variables, and offers insights into various categories and average workload hours, facilitating a deeper understanding of absenteeism patterns. It provides a comprehensive detail of each employee registered on the dataset for the organization.